Updated both Global and passport to integrate
Had to implement tx.origin to ensure access control is correct. I understand that there might be some concerns regarding tx.origin due tx.origin attacks. But based on this scenario, i can't see a way for that attack to be feasible in disrupting our network because the start point will always be a country network which is just one account. Please double check thou. Thanks.